# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger",
    "description": "Swagger",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/actor": {
      "get": {
        "tags": ["Actor"],
        "summary": "Get all Actor",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Actor"
            }
          }
        }
      },
      "post": {
        "tags": ["Actor"],
        "description": "Create new Actor",
        "parameters": [
          {
            "name": "actor",
            "in": "body",
            "description": "Actor that we want to create",
            "schema": {
              "$ref": "#/definitions/Actor"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New Actor is created",
            "schema": {
              "$ref": "#/definitions/Actor"
            }
          }
        }
      }
    },
...
    "/customer": {
      "get": {
        "tags": ["Customer"],
        "summary": "Get all Customer",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          }
        }
      },
      "post": {
        "tags": ["Customer"],
        "description": "Create new Customer",
        "parameters": [
          {
            "name": "Customer",
            "in": "body",
            "description": "Customer that we want to create",
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New Customer is created",
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Actor": {
      "required": ["first_name", "last_name"],
      "properties": {
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        }
      }
    },
    "Customer": {
      "required": [ "store_id", "first_name", "last_name", "address_id", "active"],
      "properties": {
        "store_id": { 
          "type": "number"
        },
        "first_name": { 
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "address_id": {
          "type": "number"
        },
        "active": {
        }
      }
    }
  },
...
}